import packet handler inter-stage queues and "assembly"
The "Assembly" defines the inter-stage interface used for shuttling
packets through the packet handler.  All inter-stage and shared-resource
communication and access is mediated through the assembly.

The "Packet" is the primary means of referencing a packet traveling through
the packet handler.

This commit doesn't actually include any of the processes which service
the inter-stage queues.