add and use two_way_queue
two_way_queue is a special type of queue used for sending packets
from datapath to mgmt.  Packets are send zero-copy (i.e. owned) to
mgmt, but only exposed to mgmt as a reference.  Mgmt must copy
the packet if it wants to hold it long-term.  When mgmt is done with
the original packet, it is destructed into a buffer and returned
to datapath.

This means that datapath retains ownership of all datapath packets.  This
patch thus also ensures that mgmt is no longer interacting with the buffer
stack, in preparation for making the buffer stack local per datapath lane.

This also means that mgmt and datapath now only communicate exclusively
through queues, and nonblocking shared-memory operations (e.g. RCU).